requirements.txt
anyio==4.1.0
APScheduler==3.6.3
cachetools==4.2.2
certifi==2023.11.17
emoji==0.5.4
ephem==4.1.5
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
install==1.3.5
python-telegram-bot==13.15
pytz==2023.3.post1
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
tornado==6.1
tzdata==2023.3
tzlocal==5.2
